refactor: component generation module
