+++ src/context/index.ts
@@ -20,7 +20,7 @@ const logs = new Alerts(chalk)
 const status = new Status(chalk, logs)
 const settings = new Settings()
 const config = new Config(_inquirer, chalk, settings)
-const components = new Components(logs, settings)
+const components = new Components(logs, chalk, settings)
 const commands = new Commands(status, settings, components, config)
 const shell = new Shell(Yargs, chalk, commands)
