+++ src/modules/templates/component.ts
